fix move logic to custom_batch_logger
